refactor: modularize horary engine
